backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
chromadb
pypdf
sentence-transformers
python-dotenv
tiktoken
